\name{plot.att}
\alias{plot.att}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Attack plot}
\description{Shows the changes in the network structure in 3 properties(density,average path length, and number of components) after each node removal}
\usage{
plot.att(attackscores, attacktype, centmeas)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{attackscores}{dataframe with results from the chosen attack and random attack}
}
\details{
}
\value{
plots with the network properties results after each node removal
}
\references{
%% ~put references to the literature/web site here ~
}
\author{Daniel Castro}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
